Aquatech-Bot-Backend/requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
tiktoken
PyPDF2
langchain-community
fastapi
uvicorn
httpx
